pyproject.toml
[tool.poetry]
name = "tool"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yaml2csv = 'tool.yaml2csv:all'
